fix(fonts): clip text backgrounds, keep FONT_W/H signed, fill per line
Review follow-up to 8c4e1ef. Four fixes, no change to the glyph
positioning arithmetic.

1. Opaque text backgrounds escaped the clip rectangle. gfx_rect() fills
   through gfx_hline_raw(), which clamps to the target but has no
   clip_active test; the old renderer painted background pixels through
   gfx_setpixel(), which honours it. A clipped opaque string therefore
   painted a full line-height band past its window edge -- reachable via
   apps/app_files.c:118 (Files->Rename), where gfx_set_clip() is the
   only thing bounding a filename of up to 255 chars. Added a local
   clip-aware text_bg_fill() rather than touching gfx_hline_raw(), whose
   callers are the entire tree.

2. FONT_W/FONT_H silently became unsigned when they were pointed at the
   u32 globals, so `int - FONT_H` promoted to unsigned and a negative
   intermediate wrapped. Live at apps/app_editor.c:210: at
   GFX_FONT_SCALE 2, (28 - 34)/2 evaluated to 2147483645 and drew the
   status bar text two billion pixels down. Both macros now cast to i32,
   restoring the original signed semantics.

3. gfx_char() again substitutes '?' for characters outside 32..126, as
   the old renderer did. apps/app_browser.c:167 feeds it arbitrary page
   bytes, and gfx_str_clipped() routes every character through it.

4. draw_text() fills one background band per line instead of one per
   character. The per-character fill erased the previous glyph's
   rightmost antialiased column wherever ink outruns the advance (live
   at apps/app_pkgmgr.c:139, FONT_H2 with opaque COL_SURFACE, where the
   JBM 20px face reaches bearing_x+w=13 against advance=12). Also fixes
   the dirty rect being a pixel short on the right, and stops two
   gfx_dirty calls per character overrunning MAX_DIRTY after ~16
   characters of opaque text.

+++ gui/gui.h
@@ -163,9 +163,16 @@ typedef enum {
 
 /* Active font metrics. Variables, not constants -- font_set_family() updates
  * them. No call site uses these in a constant expression, so every existing
- * user keeps compiling unchanged. */
-#define FONT_W  GFX_FONT_W
-#define FONT_H  GFX_FONT_H
+ * user keeps compiling unchanged.
+ *
+ * The (i32) cast is load-bearing, not decoration. These were the plain-int
+ * literals 8 and 13; GFX_FONT_W/H are u32. Without the cast, any surrounding
+ * `int - FONT_H` promotes to unsigned, so a negative intermediate wraps to
+ * ~4.29e9 and a following /2 yields ~2.1e9 -- e.g. centring a text row inside
+ * a bar shorter than the line height. Casting back to i32 preserves the exact
+ * signed arithmetic every call site was written against. */
+#define FONT_W  ((i32)GFX_FONT_W)
+#define FONT_H  ((i32)GFX_FONT_H)
 
 void gfx_str_ex(i32 x, i32 y, const char *s, u32 fg, u32 bg, font_size_t size);
 void gfx_str_centered_ex(i32 x, i32 y, i32 w, const char *s, u32 fg, u32 bg, font_size_t size);
